% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallel}
\alias{parallel}
\title{Evaluate R expressions in parallel}
\usage{
parallel(...)
}
\arguments{
\item{...}{The expressions to evaluate. Assignments assign to
the parent environment, see examples below.}
}
\value{
Nothing, currently.
}
\description{
Evaluate R expressions in parallel
}
\examples{
\dontrun{
x <- 1:10
parallel(
  y1 <- x + 1,
  y2 <- x * 2
)
y1
y2
}
}

