% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{show_cluster}
\alias{show_cluster}
\title{Show a brief summary of the current cluster}
\usage{
show_cluster()
}
\value{
An invisible list, with arguments:
  \item{num_workers}{The number of worker nodes}
  \item{state}{The state of the cluster, a character vector with
    one entry for each worker. The state of the worker is either
    \sQuote{free} or \sQuote{busy}.}
}
\description{
It also returns a machine readable list, invisibly.
The return value can be used to test if the cluster is set up:
if not, then the \code{num_workers} entry is set to zero.
}
\seealso{
Other cluster.management: \code{\link{start_cluster}},
  \code{\link{stop_cluster}}
}

