% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{start_cluster}
\alias{start_cluster}
\title{Start the cluster}
\usage{
start_cluster(workers = NULL)
}
\arguments{
\item{workers}{Number of worker processes. If \code{NULL}, then
the number of (logical) cores of the machine is used.}
}
\description{
Start the cluster
}
\seealso{
Other cluster.management: \code{\link{show_cluster}},
  \code{\link{stop_cluster}}
}

