% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{stop_cluster}
\alias{stop_cluster}
\title{Stop the cluster}
\usage{
stop_cluster()
}
\description{
Stop the cluster
}
\seealso{
Other cluster.management: \code{\link{cluster}},
  \code{\link{start_cluster}}
}

